feat(table): mobile responsive
